src/consumer/requirements.txt
confluent_kafka
python-dotenv
numpy
scikit-learn
xgboost
lightgbm
mlflow
pyyaml
